fix: copy text natively
