+++ src/components/SlateEditor/plugins/DND/index.js
@@ -1,5 +1,7 @@
-import { findNode } from 'slate-react';
+import { findNode, getEventRange } from 'slate-react';
+import { Text } from 'slate';
 import onDrop from './onDrop';
+import { getTopNode } from './utils';
 
 function onDragOver(event, editor, next) {
   event.stopPropagation();
@@ -8,6 +10,11 @@ function onDragOver(event, editor, next) {
 
 function onDragStart(event, editor, next) {
   const dragSource = findNode(event.target, editor);
+  const { type } = getTopNode(dragSource, editor);
+  if (Text.isText(dragSource) && type !== 'table' && !type.includes('list')) {
+    // just copy the text natively
+    return next();
+  }
   event.dataTransfer.setData('text/nodeKey', dragSource.key);
 }
